requirements.txt
supabase
python-dotenv
streamlit
pygetwindow
pynput
pyrect
pandas 
plotly
